camera behaviour added
2 scripts were added:
FirstPersonScript - a script responsible for controlling the camera rotation according to user input
CameraHolder - to avoid bugs with including a camera into a rigidbody, a camera holder is usually used to maintain the camera inside the player's rigidbody. this script just assures that the camera will always be in it's position.

All can be checked in the scene: "/Assets/Scenes/Testing/CameraMovementTesting"